+++ sockeye/loss.py
@@ -14,6 +14,7 @@
 """
 Functions to generate loss symbols for sequence-to-sequence models.
 """
+import logging
 from abc import ABC, abstractmethod
 from typing import List, Optional
 
@@ -22,7 +23,8 @@ from mxnet.metric import EvalMetric
 
 from . import config
 from . import constants as C
-from . import utils
+
+logger = logging.getLogger(__name__)
 
 
 class LossConfig(config.Config):
@@ -31,20 +33,20 @@ class LossConfig(config.Config):
 
     :param name: Loss name.
     :param vocab_size: Target vocab size.
-    :param normalize: Whether to normalize loss value.
-    :param smoothed_cross_entropy_alpha: Smoothing value for smoothed-cross-entropy loss.
+    :param normalization_type: How to normalize the loss.
+    :param label_smoothing: Optional smoothing constant for label smoothing.
     """
 
     def __init__(self,
                  name: str,
                  vocab_size: int,
-                 normalize: bool,
-                 smoothed_cross_entropy_alpha: float = 0.0) -> None:
+                 normalization_type: bool,
+                 label_smoothing: float = 0.0) -> None:
         super().__init__()
         self.name = name
         self.vocab_size = vocab_size
-        self.normalize = normalize
-        self.smoothed_cross_entropy_alpha = smoothed_cross_entropy_alpha
+        self.normalization_type = normalization_type
+        self.label_smoothing = label_smoothing
 
 
 def get_loss(loss_config: LossConfig) -> 'Loss':
@@ -55,21 +57,8 @@ def get_loss(loss_config: LossConfig) -> 'Loss':
     """
     if loss_config.name == C.CROSS_ENTROPY:
         return CrossEntropyLoss(loss_config)
-    elif loss_config.name == C.SMOOTHED_CROSS_ENTROPY:
-        return SmoothedCrossEntropyLoss(loss_config)
     else:
         raise ValueError("unknown loss name: %s" % loss_config.name)
-
-
-def _normalize(loss: mx.sym.Symbol, labels: mx.sym.Symbol):
-    """
-    Normalize loss by the number of non-PAD tokens.
-
-    :param loss: A loss value for each label.
-    :param labels: A label for each loss entry (potentially containing PAD tokens).
-    :return: The normalized loss.
-    """
-    return mx.sym.broadcast_div(loss, mx.sym.sum(labels != C.PAD_ID))
 
 
 class Loss(ABC):
@@ -107,6 +96,8 @@ class CrossEntropyLoss(Loss):
     """
 
     def __init__(self, loss_config: LossConfig) -> None:
+        logger.info("Loss: CrossEntropy(normalization_type=%s, label_smoothing=%s)",
+                    loss_config.normalization_type, loss_config.label_smoothing)
         self.loss_config = loss_config
 
     def get_loss(self, logits: mx.sym.Symbol, labels: mx.sym.Symbol) -> List[mx.sym.Symbol]:
@@ -117,15 +108,18 @@ class CrossEntropyLoss(Loss):
         :param labels: Shape: (batch_size * target_seq_len,).
         :return: List of loss symbol.
         """
-        if self.loss_config.normalize:
+        if self.loss_config.normalization_type == C.LOSS_NORM_VALID:
             normalization = "valid"
-        else:
+        elif self.loss_config.normalization_type == C.LOSS_NORM_BATCH:
             normalization = "null"
+        else:
+            raise ValueError("Unknown loss normalization type: %s" % self.loss_config.normalization_type)
         return [mx.sym.SoftmaxOutput(data=logits,
                                      label=labels,
                                      ignore_label=C.PAD_ID,
                                      use_ignore=True,
                                      normalization=normalization,
+                                     smooth_alpha=self.loss_config.label_smoothing,
                                      name=C.SOFTMAX_NAME)]
 
     def create_metric(self) -> "CrossEntropyMetric":
@@ -169,88 +163,3 @@ class CrossEntropyMetric(EvalMetric):
                 # NOTE: This is different from MXNet's metrics
                 self.num_inst += batch_size
             self.sum_metric += mx.nd.sum(loss).asscalar()
-
-
-class SmoothedCrossEntropyLoss(Loss):
-    """
-    Computes a smoothed cross-entropy loss. Smoothing is defined by alpha which indicates the
-    amount of probability mass subtracted from the true label probability (1-alpha).
-    Alpha is then uniformly distributed across other labels.
-    If normalize is True, the gradient by dividing by the number of non-PAD tokens.
-
-    :param loss_config: Loss configuration.
-    """
-
-    def __init__(self, loss_config: LossConfig) -> None:
-        self.loss_config = loss_config
-        utils.check_condition(self.loss_config.smoothed_cross_entropy_alpha >= 0,
-                              "alpha for smoothed loss must be >= 0")
-
-    def get_loss(self, logits: mx.sym.Symbol, labels: mx.sym.Symbol) -> List[mx.sym.Symbol]:
-        """
-        Returns loss and softmax output symbols given logits and integer-coded labels.
-
-        :param logits: Shape: (batch_size * target_seq_len, target_vocab_size).
-        :param labels: Shape: (batch_size * target_seq_len,).
-        :return: List of loss and softmax output symbols.
-        """
-        probs = mx.sym.softmax(data=logits)
-
-        on_value = 1.0 - self.loss_config.smoothed_cross_entropy_alpha
-        off_value = self.loss_config.smoothed_cross_entropy_alpha / (self.loss_config.vocab_size - 1.0)
-        cross_entropy = mx.sym.one_hot(indices=mx.sym.cast(data=labels, dtype='int32'),
-                                       depth=self.loss_config.vocab_size,
-                                       on_value=on_value,
-                                       off_value=off_value)
-
-        # zero out pad symbols (0)
-        cross_entropy = mx.sym.where(labels, cross_entropy, mx.sym.zeros((0, self.loss_config.vocab_size)))
-
-        # compute cross_entropy
-        cross_entropy = cross_entropy * (-mx.sym.log(data=probs + 1e-10))
-        cross_entropy = mx.sym.sum(data=cross_entropy, axis=1)
-
-        if self.loss_config.normalize:
-            cross_entropy = _normalize(cross_entropy, labels)
-
-        cross_entropy = mx.sym.MakeLoss(cross_entropy, name=C.SMOOTHED_CROSS_ENTROPY)
-        probs = mx.sym.BlockGrad(probs, name=C.SOFTMAX_NAME)
-        return [cross_entropy, probs]
-
-    def create_metric(self) -> "SmoothedCrossEntropyMetric":
-        return SmoothedCrossEntropyMetric(self.loss_config)
-
-
-class SmoothedCrossEntropyMetric(EvalMetric):
-    """
-    Metric wrapper for smoothed cross entropy loss.  Since SCE already returns loss values during
-    training, this class simply sums the results.
-
-    :param loss_config: The configuration used for the corresponding loss.
-    :param name: Name of this metric instance for display.
-    :param output_names: Name of predictions that should be used when updating with update_dict.
-    :param output_labels: Name of labels that should be used when updating with update_dict.
-    """
-
-    def __init__(self,
-                 loss_config: LossConfig,
-                 name: str = C.SMOOTHED_CROSS_ENTROPY,
-                 output_names: Optional[List[str]] = None,
-                 label_names: Optional[List[str]] = None) -> None:
-        super().__init__(name, output_names=output_names, label_names=label_names)
-        self.loss_config = loss_config
-
-    def update(self, labels, preds):
-        # Take cross entropy values only
-        sces = preds[::2]
-        for label, sce in zip(labels, sces):
-            batch_size = label.shape[0]
-            # SCE is pre-computed, so just sum
-            self.sum_metric += sce.asnumpy().sum()
-            # Only scale if loss is not already normalized
-            if self.loss_config.normalize:
-                self.num_inst += 1
-            else:
-                # When not normalizing, we divide by the batch size (number of sequences)
-                # NOTE: This is different from MXNet's metrics
-                self.num_inst += batch_size