+++ CHANGELOG.md
@@ -10,6 +10,22 @@ Note that Sockeye has checks in place to not translate with an old model that wa
 
 For each item we will potentially have subsections for: _Added_, _Changed_, _Removed_, _Deprecated_, and _Fixed_.
 
+## [1.10.0]
+### Changed
+ - Updated MXNet dependency to 0.12 (w/ MKL support by default).
+ - Changed `--smoothed-cross-entropy-alpha` to `--label-smoothing`.
+ Label smoothing should now require signficantly less memory due to its addition to MXNet's `SoftmaxOutput` operator.
+ - `--weight-normalization` now applies not only to convolutional weight matrices, but to output layers of all decoders.
+ It is also independent of weight tying.
+
+### Added
+ - `--loss-normalization-type`. Added a new flag to control loss normalization. New default is to normalize
+ by the number of valid, non-PAD tokens instead of the batch size.
+ 
+### Removed
+ - `--smoothed-cross-entropy-alpha` argument. See above.
+ - `--normalize-loss` argument. See above.
+
 ## [1.9.0]
 ### Added
  - Batch decoding. New options for the translate CLI: ``--batch-size`` and ``--chunk-size``. Translator.translate()